Replace the original AWS job store and trigger testing (https://github.com/DataBiosphere/toil/pull/3153#issuecomment-678570705)
- Also fixed a few issues with boto3 integration
